% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simMatch.R
\name{plot.svR_matches}
\alias{plot.svR_matches}
\title{plot method for detailed return of \link{simMatches}}
\usage{
\method{plot}{svR_matches}(x)
}
\description{
plot method for detailed return of \link{simMatches}
}

