% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{plot.svR_sets}
\alias{plot.svR_sets}
\title{plot method for detailed return of \link{simSets}}
\usage{
\method{plot}{svR_sets}(x)
}
\description{
plot method for detailed return of \link{simSets}
}

