% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{print.svR_game}
\alias{print.svR_game}
\title{print method for detailed return of \link{simGame}}
\usage{
\method{print}{svR_game}(x, ...)
}
\description{
print method for detailed return of \link{simGame}
}

