% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{print.svR_games}
\alias{print.svR_games}
\title{print method for detailed return of \link{simGames}}
\usage{
\method{print}{svR_games}(x)
}
\description{
print method for detailed return of \link{simGames}
}

