% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simMatch.R
\name{print.svR_match}
\alias{print.svR_match}
\title{print method for detailed return of \link{simMatch}}
\usage{
\method{print}{svR_match}(x)
}
\description{
print method for detailed return of \link{simMatch}
}

