% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simMatch.R
\name{print.svR_matches}
\alias{print.svR_matches}
\title{print method for detailed return of \link{simMatches}}
\usage{
\method{print}{svR_matches}(x)
}
\description{
print method for detailed return of \link{simMatches}
}

