% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{print.svR_set}
\alias{print.svR_set}
\title{print method for detailed return of \link{simSet}}
\usage{
\method{print}{svR_set}(x)
}
\description{
print method for detailed return of \link{simSet}
}

