% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{print.svR_sets}
\alias{print.svR_sets}
\title{print method for detailed return of \link{simSets}}
\usage{
\method{print}{svR_sets}(x)
}
\description{
print method for detailed return of \link{simSets}
}

