% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/servevolleyR.R
\docType{package}
\name{servevolleyR}
\alias{servevolleyR}
\alias{servevolleyR-package}
\title{servevolleyR}
\description{
Package for simulating tennis games, tiebreaks, sets, and matches
}
\author{
Thomas Heslop
}

