% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simDf.R
\name{simDf}
\alias{simDf}
\title{simDf}
\usage{
simDf(x, ...)
}
\arguments{
\item{...}{recommend adding \link{plyr}'s .progress when using \link{simGames},
\link{simSets} or \link{simMatches} to view progress at processing results.}

\item{sim}{detailed list returned by one of the following functions \link{simGame}
\link{simGames}, \link{simSet}, \link{simSets}, \link{simMatch} or \link{simMatches}}
}
\description{
convert simulated games, sets and matches into useful dataframes
for further analysis
}

