% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{simGame}
\alias{simGame}
\title{simGame}
\usage{
simGame(p, p2 = NULL, firstServe = NULL, detail = FALSE, player = "A")
}
\arguments{
\item{p}{probability of server winning point on their first serve}

\item{p2}{probability of server winning point on their second serve}

\item{firstServe}{probability of first serve being in}

\item{detail}{return detailed output, default of FALSE returns 1 or 0, 1 if
server wins, 0 otherwise}

\item{player}{player name, default is "A"}
}
\value{
1 (if server wins), 0 (if returner wins).  If the parameter \strong{detail}
is set to TRUE, then the function will return a detailed list about the simulated
game, see \link{simGames} for more about simulating many games for a single server.
This detailed list can be converted into a dataframe using \link{simDf}
}
\description{
simulate probability of a server winning a game given the probability
they win a point on their serve
}
\details{
a couple of additional parameters can be supplied: the probability of
the server winning a point on their second serve and probability of their first
serve going in.
}

