% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{simGames}
\alias{simGames}
\title{simulate many Games}
\usage{
simGames(n = 1000, p, p2 = NULL, firstServe = NULL, player = "A",
  .progress = "none")
}
\arguments{
\item{n}{number of games to simulate (default, 1000)}

\item{p}{probability of server winning a point on the first serve}

\item{p2}{probability of server winning a point on their second serve}

\item{firstServe}{probability of first serve being in}

\item{player}{player name, default is "A"}

\item{.progress}{\link{plyr}'s progress bar}
}
\description{
simulate many service games for a single server
}
\details{
adding a parameter ("text", or "time") to the .progress argument will
cause two progress bars to show up, one after another, the first charts the
simulation of many games, while the second processes these results into something
more useful for users.
}

