% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simMatch.R
\name{simMatch}
\alias{simMatch}
\title{simMatch}
\usage{
simMatch(pA, pB, sets = c(3, 5), tiebreaks = TRUE,
  finalSetTiebreak = FALSE, players = c("A", "B"), detail = FALSE,
  p2A = NULL, firstServeA = NULL, p2B = NULL, firstServeB = NULL)
}
\arguments{
\item{pA}{probability of player A winning point on their serve}

\item{pB}{probability of player B winning point on their serve}

\item{sets}{number of sets to be played}

\item{tiebreaks}{play tie break at 6-6, or keep playing}

\item{finalSetTiebreak}{play tie break at 6-6 in final set}

\item{players}{player names, vector of length 2}

\item{detail}{return detailed data for the Match, default FALSE}

\item{p2A}{probability of player A winning point on their second serve}

\item{firstServeA}{probability of player A getting their first serve in}

\item{p2B}{probability of player B winning point on their second serve}

\item{firstServeB}{probability of player B getting their first serve in}
}
\value{
1 (if player A wins) or 0 (if player B wins). If the parameter \strong{detail}
is set to TRUE, then the function will return a detailed list about the simulated
match, including data about the indvidual sets and games in the match.  This list can
be converted to a dataframe using \link{simDf}. See \link{simMatches} for simulating
many sets between the two players.
}
\description{
simulate a match between two players, given the probability of each
player winning a point on their serve
}
\details{
minimum input required is the probability that each player will win
a point on their serve, however additional parameters include probability
of each player winning a point on their second serve, the probability of a
players first serve being in, number of sets whether they play a tiebreak
}

