% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simPoint.R
\name{simPoint}
\alias{simPoint}
\title{simPoint}
\usage{
simPoint(p, p2 = NULL, firstServe = NULL)
}
\arguments{
\item{p}{probability of server winning point on their first serve}

\item{p2}{probability of server winning point on their second serve}

\item{firstServe}{probability of first serve being in}
}
\value{
1 (if server wins), 0 (if returner wins)
}
\description{
simulate probability of server winning a point
}

