% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{simSet}
\alias{simSet}
\title{simulate a set between two players}
\usage{
simSet(pA, pB, playTiebreak = TRUE, players = c("A", "B"), detail = FALSE,
  p2A = NULL, firstServeA = NULL, p2B = NULL, firstServeB = NULL)
}
\arguments{
\item{pA}{probability of player A winning point on their first serve}

\item{pB}{probability of player B winning point on their first serve}

\item{playTiebreak}{logical, play tiebreak at 6 games all}

\item{players}{player names, vector of length 2}

\item{detail}{return detailed data for set, default FALSE}

\item{p2A}{probability of player A winning point on their second serve}

\item{firstServeA}{probability of player A getting their first serve in}

\item{p2B}{probability of player B winning point on their second serve}

\item{firstServeB}{probability of player B getting their first serve in}
}
\description{
simulate a set between two players given a number of parameters,
that include probability of each player winning a point on their first serve,
on their second serve, and the probability that their first serve will go in.
}
\details{
minimum input required is the probability that each player will win
a point on their serve, however additional parameters include probability
of each player winning a point on their second serve, the probability of a
players first serve being in, and whether they play a tiebreak
}

