% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{simSets}
\alias{simSets}
\title{simSets}
\usage{
simSets(n = 1000, pA, pB, playTiebreak = TRUE, players = c("A", "B"),
  p2A = NULL, firstServeA = NULL, p2B = NULL, firstServeB = NULL,
  .progress = "none")
}
\arguments{
\item{n}{number of simulations, default of 1000}

\item{pA}{probability of player A winning point on their first serve}

\item{pB}{probability of player B winning point on their first serve}

\item{playTiebreak}{logical, play tiebreak at 6 games all}

\item{players}{player names, vector of length 2}

\item{p2A}{probability of player A winning point on their second serve}

\item{firstServeA}{probability of player A getting their first serve in}

\item{p2B}{probability of player B winning point on their second serve}

\item{firstServeB}{probability of player B getting their first serve in}

\item{.progress}{\link{plyr}'s progress bar}
}
\value{
The function returns a large list, which can be printed, summarised, or
plotted. It can also be converted to a dataframe using \link{simDf}, which contains
data about the simulated sets, and games, within each simulation.
}
\description{
simulate many sets between two players, given the probability of
each player winning a point on their serve
}
\details{
minimum input required is the probability that each player will win
a point on their serve, however additional parameters include probability
of each player winning a point on their second serve, the probability of a
players first serve being in, and whether they play a tiebreak.

Adding a parameter ("text", or "time") to the .progress argument will
cause two progress bars to show up, one after another, the first charts the
simulation of many games, while the second processes these results into something
more useful for users.
}

