% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTiebreak.R
\name{simTiebreak}
\alias{simTiebreak}
\title{simTiebreak}
\usage{
simTiebreak(pA, pB, p2A = NULL, firstServeA = NULL, p2B = NULL,
  firstServeB = NULL)
}
\arguments{
\item{pA}{probability of player A winning point on their serve}

\item{pB}{probability of player B winning point on their serve}

\item{p2A}{probability of player A winning point on their second serve}

\item{firstServeA}{probability of player A getting their first serve in}

\item{p2B}{probability of player B winning point on their second serve}

\item{firstServeB}{probability of player B getting their first serve in}
}
\description{
simulate a tiebreak between two players
}

