% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{summary.svR_game}
\alias{summary.svR_game}
\title{summary method for detailed return of \link{simGame}}
\usage{
\method{summary}{svR_game}(x, ...)
}
\description{
summary method for detailed return of \link{simGame}
}

