% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{summary.svR_games}
\alias{summary.svR_games}
\title{summary method for detailed return of \link{simGames}}
\usage{
\method{summary}{svR_games}(x)
}
\description{
summary method for detailed return of \link{simGames}
}

