% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simMatch.R
\name{summary.svR_match}
\alias{summary.svR_match}
\title{summary method for detailed return of \link{simMatch}}
\usage{
\method{summary}{svR_match}(x)
}
\description{
summary method for detailed return of \link{simMatch}
}

