% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simMatch.R
\name{summary.svR_matches}
\alias{summary.svR_matches}
\title{summary method for detailed return of \link{simMatches}}
\usage{
\method{summary}{svR_matches}(x)
}
\description{
summary method for detailed return of \link{simMatches}
}

