% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{summary.svR_set}
\alias{summary.svR_set}
\title{summary method for detailed return of \link{simSet}}
\usage{
\method{summary}{svR_set}(x)
}
\description{
summary method for detailed return of \link{simSet}
}

