% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simSet.R
\name{summary.svR_sets}
\alias{summary.svR_sets}
\title{summary method for detailed return of \link{simSets}}
\usage{
\method{summary}{svR_sets}(x)
}
\description{
summary method for detailed return of \link{simSets}
}

