% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simGame.R
\name{simGame}
\alias{simGame}
\title{Simulate a server winning a game}
\usage{
simGame(p, p2 = NULL, firstServe = NULL, detail = FALSE, player = "A")
}
\arguments{
\item{p}{probability of server winning point on their first serve}

\item{p2}{probability of server winning point on their second serve}

\item{firstServe}{probability of first serve being in}

\item{detail}{return detailed output, default of FALSE returns 1 or 0, 1 if
server wins, 0 otherwise}

\item{player}{player name, default is "A"}
}
\description{
simulate probability of a server winning a game given a number of
parameters, that include probability of server winning a point on their first
serve, their second serve, and probability of their first serve going in.
}
\details{
if the parameter \strong{detail} is set to TRUE, then the function
will return a detailed list about the simulated game, see \link{simGames} for
more about simulating many games for a single server
}

